Add site: Justin Lin
To add your site to the webring, you must be a current student or an alum of the undergraduate Software Engineering program at the University of Waterloo (Waterloo, ON, Canada).

Full Name: Justin Lin

Cohort Year: 2027

Full website URL: https://justin-lin.com

So we know you're human:

LinkedIn or a similar profile: https://www.linkedin.com/in/justin-lin-905/

Don't forget to also add your entry to the end of the allSites array in the se-webring.js file (name, year, personal website), and to add the add site label to your pull request.